Release v1.0.2 (#95)
## Background
Several small scale bug-fixes will be released as a final version for
users to enjoy before having to download updated commec database files
for the upcoming Control Lists release.

## Changes
* Fixes erroneous "Matches ." Rationale text that could occur under some
Biorisk output conditions.
* Fixed an issue where Benign clears were being applied to all regions
of multiple region regulated hit.
* Taxonomic information is sorted in the json output
* A maximum of 10 entries of non-regulated are shown in the json output
in the case of a mixed hit result.
* E-values are now included for each individual taxonomic information
per hit
* Updated the description for each hit to include a rationale-like
output depending on case, rather than just the description of the first
entry.
* Rationale texts are added as a column to `commec flag` output csv.
* Additional support for up to 255 characters of filename size, for
those abhorrently long filenames.
* More appropriate error handling depending on whether or not --Resume,
--Force are used and the presence of pre-existing output files.
* Certain temporary output files - derived from the output of Blastx,
Blastn, hmmscan, for example are now kept in the event of a failed run.
* Commec now differentiates between a query having no hits (with all
steps being run fully) and no hits, skipped (when taxonomy was skipped
for example with --skip-tx) in its output status'

### New features
* Adds additional Taxonomic info for every hit, to aide in human
investigation of any particular mixture of regulated and non-regulated
results.

### Refactoring
* Minor tidying of the check regulated pathogens function.

---------

+++ commec/tools/search_handler.py
@@ -65,6 +65,7 @@ class SearchHandler(ABC):
         self.threads = kwargs.get('threads', 1)
         self.force = kwargs.get('force', False)
         self.arguments_dictionary = {}
+        self.successful = True
 
         # Only validate database files if we actually intend on using them
         if not self.should_use_existing_output:
@@ -179,6 +180,8 @@ class SearchHandler(ABC):
         """
         Run a command using subprocess.run, piping stdout and stderr to `out_file`.
         """
+        self.successful = False
+
         logger.debug("SUBPROCESS: %s", " ".join(command))
         logger.debug(" ".join(command), extra = {"no_prefix":True,"cap":True})
 
@@ -198,7 +201,9 @@ class SearchHandler(ABC):
                     f"subprocess.run of command '{command_str}' encountered error."
                     f" Check {out_file} for logs."
                 )
+            
+        self.successful = True
 
     def __del__(self):
-        if os.path.exists(self.temp_log_file):
+        if os.path.exists(self.temp_log_file) and self.successful:
             os.remove(self.temp_log_file)
